+++ modules/secure-serverless-harness/private_service_connect.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,16 @@
  * limitations under the License.
  */
 
-terraform {
-  required_version = ">= 0.13"
-  required_providers {
-    terracurl = {
-      source  = "devops-rob/terracurl"
-      version = "~> 1.0"
-    }
-    google = {
-      source  = "hashicorp/google"
-      version = "< 5.0"
-    }
-  }
-  provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-cloud-run:job-exec/v0.7.0"
-  }
+module "private_service_connect" {
+  for_each = module.network
+
+  source                     = "terraform-google-modules/network/google//modules/private-service-connect"
+  version                    = "~> 6.0"
+  project_id                 = each.value.project_id
+  network_self_link          = each.value.network_self_link
+  private_service_connect_ip = var.private_service_connect_ip
+  forwarding_rule_target     = "vpc-sc"
+  depends_on = [
+    time_sleep.wait_180_seconds
+  ]
 }