+++ modules/secure-serverless-net/outputs.tf
@@ -29,11 +29,6 @@ output "cloud_services_sa" {
   description = "Google APIs service agent."
 }
 
-output "run_identity_services_sa" {
-  value       = google_project_service_identity.run_identity_sa.email
-  description = "Google APIs Service Agent to Cloud Run Service."
-}
-
 output "subnet_name" {
   value       = local.subnet_name
   description = "The name of the sub-network used to create VPC Connector."